+++ tsup.config.ts
@@ -26,7 +26,6 @@ export default defineConfig([
       "store/QueueResource": "src/store/queueResource.ts",
       "store/Log": "src/store/log.ts",
       "store/ProcessLifecycle": "src/store/processLifecycle.ts",
-      "store/ProcessExecution": "src/store/processExecution.ts",
       Query: "src/Query.ts",
       ResourceConfigure: "src/ResourceConfigure.ts",
       RuntimeStorage: "src/RuntimeStorage.ts",